Refactor position forwarding
