+++ src/main/java/org/traccar/BasePipelineFactory.java
@@ -140,7 +140,7 @@ public abstract class BasePipelineFactory extends ChannelInitializer<Channel> {
                 CopyAttributesHandler.class,
                 EngineHoursHandler.class,
                 ComputedAttributesHandler.class,
-                WebDataHandler.class,
+                PositionForwardingHandler.class,
                 DefaultDataHandler.class,
                 MediaEventHandler.class,
                 CommandResultEventHandler.class,
