+++ src/main/java/org/traccar/forward/ResultHandler.java
@@ -15,6 +15,6 @@
  */
 package org.traccar.forward;
 
-public interface EventForwarder {
-    void forward(EventData eventData, ResultHandler resultHandler);
+public interface ResultHandler {
+    void onResult(boolean success, Throwable throwable);
 }
