+++ src/test/java/org/traccar/forward/PositionForwarderUrlTest.java
@@ -1,19 +1,17 @@
-package org.traccar;
+package org.traccar.forward;
 
 import org.junit.Test;
+import org.traccar.ProtocolTest;
 import org.traccar.config.Config;
 import org.traccar.config.Keys;
 import org.traccar.model.Device;
 import org.traccar.model.Position;
-import org.traccar.session.cache.CacheManager;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class WebDataHandlerTest extends ProtocolTest {
+public class PositionForwarderUrlTest extends ProtocolTest {
 
     @Test
     public void testFormatRequest() throws Exception {
@@ -28,14 +26,16 @@ public class WebDataHandlerTest extends ProtocolTest {
         when(device.getName()).thenReturn("test");
         when(device.getUniqueId()).thenReturn("123456789012345");
         when(device.getStatus()).thenReturn(Device.STATUS_ONLINE);
-        var cacheManager = mock(CacheManager.class);
-        when(cacheManager.getObject(eq(Device.class), anyLong())).thenReturn(device);
 
-        WebDataHandler handler = new WebDataHandler(config, cacheManager, null, null, null);
+        PositionData positionData = new PositionData();
+        positionData.setPosition(position);
+        positionData.setDevice(device);
+
+        PositionForwarderUrl forwarder = new PositionForwarderUrl(config, null, null);
 
         assertEquals(
                 "http://localhost/?fixTime=1451610123000&gprmc=$GPRMC,010203.000,A,2000.0000,N,03000.0000,E,0.00,0.00,010116,,*05&name=test",
-                handler.formatRequest(position));
+                forwarder.formatRequest(positionData));
 
     }
